% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{batch_extract_metadata}
\alias{batch_extract_metadata}
\title{Batch extract metadata from multiple surveys}
\usage{
batch_extract_metadata(file_list, ...)
}
\arguments{
\item{file_list}{A list where each element is a character vector of file
paths for one survey. Or a simple character vector (one file per survey).}

\item{...}{Additional arguments passed to \code{\link[=extract_metadata]{extract_metadata()}}.}
}
\value{
A data.frame with one row per survey.
}
\description{
Batch extract metadata from multiple surveys
}
