% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{col_types}
\alias{col_types}
\title{modal column types of a list of dataframes}
\usage{
col_types(data, max_n)
}
\arguments{
\item{data}{a list of dataframes (with the expectation a single row per dataframe)}

\item{n_max}{number of dataframes to sample from the list. If n_max exceeds the length of the list,
the full set is analyzed.}
}
\value{
a dataframe
}
\description{
\code{col_types} returns the modal column types of a list of dataframes. To reduce computation time,
dataframes can be sampled from the supplied list.
}
