% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{dataframe_to_variables_list}
\alias{dataframe_to_variables_list}
\title{Convert long-format data.frame back to nested variable list}
\usage{
dataframe_to_variables_list(df)
}
\description{
Convert long-format data.frame back to nested variable list
}
\keyword{internal}
