% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_text_pdf}
\alias{extract_text_pdf}
\title{Extract text from PDF}
\usage{
extract_text_pdf(file_path, use_ocr = FALSE)
}
\arguments{
\item{file_path}{Character. Path to PDF file.}

\item{use_ocr}{Logical. If TRUE and pdftools returns little text,
attempt OCR with tesseract.}
}
\value{
Character string.
}
\description{
Extract text from PDF
}
\keyword{internal}
