% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_to_base64}
\alias{file_to_base64}
\title{Encode a file as base64}
\usage{
file_to_base64(file_path)
}
\arguments{
\item{file_path}{Character. Path to file.}
}
\value{
Character string of base64-encoded content.
}
\description{
Encode a file as base64
}
\keyword{internal}
