% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_metadata}
\alias{format_metadata}
\title{flatten metadata response into a dataframe}
\usage{
format_metadata(meta, df)
}
\arguments{
\item{meta}{A metadata response object (represetned by an r list) from the '/catalog/{IDNo}/variables/{varId}' endpoint}

\item{df}{an empty dataframe that is recursively appended}
}
\value{
a dataframe
}
\description{
\code{format_metadata} selectively flattens metadata response object into a dataframe. Atomic list items are written within dataframe columns,
and dataframe objects are nested with a column. The function is called from \code{get_var_meta}.
}
