% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gemini_variable_schema}
\alias{gemini_variable_schema}
\title{Convert variable schema to Gemini format}
\usage{
gemini_variable_schema()
}
\description{
Convert variable schema to Gemini format
}
\keyword{internal}
