% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_metadata}
\alias{get_metadata}
\title{get variable metadata for multiple surveys}
\usage{
get_metadata(base_url, ids = NULL, var_meta = TRUE)
}
\arguments{
\item{base_url}{The base url for the catalog (for IHSN this is http://catalog.ihsn.org/index.php/api/catalog/)}

\item{ids}{character vector of survey ids (idno)}

\item{var_meta}{if true returns variable metadata for queried surveys}

\item{ps}{maximum number of results}
}
\value{
a list with three dataframes. 'Result' is for successful queries; 'failed_surveys' is for failed variable queries, by survey_id;
'failed_variables' is for field variable queries, by survey and variable id.
}
\description{
\code{get_metadata} is a wrapper function that gets the metadata for a specified surveys. Operations will run in parallel if
\code{future::plan(multisession)} is set before the function call
}
