% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_survey_meta}
\alias{get_survey_meta}
\title{get survey metadata}
\usage{
get_survey_meta(base_url, survey_id)
}
\arguments{
\item{base_url}{The base url for the catelog (for IHSN this is http://catalog.ihsn.org/index.php/api/catalog/)}

\item{survey_id}{Corresponds to idno in the object returned from \code{get_surveys}.}

\item{ps}{maximum number of results}
}
\value{
a dataframe
}
\description{
\code{get_survey_meta} Queries searches for surveys published between specified dates, and returns a dataframe of matching surveys.
}
