% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_surveys}
\alias{get_surveys}
\title{survey search}
\usage{
get_surveys(base_url, min_year, max_year, ps = 10000)
}
\arguments{
\item{base_url}{The base url for the catelog (for IHSN this is http://catalog.ihsn.org/index.php/api/catalog/)}

\item{min_year}{lower bound for the year the survey was published}

\item{max_year}{upper bound for the year the survey was publishe.}

\item{ps}{maximum number of results}
}
\value{
a dataframe
}
\description{
\code{get_surveys} Queries searches for surveys published between specified dates, and returns a dataframe of matching surveys.
}
