% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_surveys_meta}
\alias{get_surveys_meta}
\title{survey search with supplementary metadata}
\usage{
get_surveys_meta(
  base_url,
  min_year = NULL,
  max_year = NULL,
  created = NULL,
  ps = 10
)
}
\arguments{
\item{base_url}{The base url for the catelog (for IHSN this is http://catalog.ihsn.org/index.php/api/catalog/)}

\item{min_year}{lower bound for the year the survey was published}

\item{max_year}{upper bound for the year the survey was published.}

\item{created}{Filter  by date of creation. Use the date format YYYY-MM-DD. E.g 2020/04/01 returns records created on and after that date.
To specify a date range, use the format 2020/04/01-2020/04/20}

\item{ps}{maximum number of results}
}
\value{
a dataframe
}
\description{
\code{get_surveys_meta} Queries searches for surveys published between specified dates, and returns a dataframe of matching surveys, with extended metadata for each survey
}
