% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_var_meta}
\alias{get_var_meta}
\title{get variable metadata}
\usage{
get_var_meta(base_url, survey_id, variable_id)
}
\arguments{
\item{base_url}{The base url for the catelog (for IHSN this is http://catalog.ihsn.org/index.php/api/catalog/)}

\item{survey_id}{Corresponds to idno in the object returned from \code{get_surveys}.}

\item{variable_id}{Corresponds to vid in the object returned from \code{get_vars}.}
}
\value{
a list with the result and error. If no error, err is set to FALSE and result is a signle row dataframe of variable metadata.
Otherwise err is TRUE and result is a dataframe of the error message.
}
\description{
\code{get_var_meta} gets the metadata for a sepcified variable_id
}
