% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabularies.R
\name{load_vocabulary}
\alias{load_vocabulary}
\title{Load a controlled vocabulary file}
\usage{
load_vocabulary(vocab_name)
}
\arguments{
\item{vocab_name}{Character. One of "cessda_topics", "ssjda_topics",
"sampling_procedures", "collection_modes".}
}
\value{
A data.frame with vocabulary entries.
}
\description{
Load a controlled vocabulary file
}
