% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.survey_evaluation}
\alias{print.survey_evaluation}
\title{Print method for evaluation results}
\usage{
\method{print}{survey_evaluation}(x, ...)
}
\arguments{
\item{x}{An survey_evaluation object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for evaluation results
}
