% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{print.survey_variables}
\alias{print.survey_variables}
\title{Print method for survey_variables}
\usage{
\method{print}{survey_variables}(x, ...)
}
\arguments{
\item{x}{A survey_variables object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for survey_variables
}
