% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reduce_survey_metadata}
\alias{reduce_survey_metadata}
\title{format and reduce a list of survey metadata to a dataframe}
\usage{
reduce_survey_metadata(data)
}
\arguments{
\item{data}{a list of dataframes (with the expectation a single row per dataframe)}
}
\value{
a dataframe
}
\description{
\code{reduce_metadata} returns a formatted dataframe of survey metadata, with a row per survey. As columns with no data that
would typically be expected as a list are of type character when after application of \code{format_metadata}, they are converted to list
based on inferring their correct type using \code{col_type}.
}
