% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{var_meta_format}
\alias{var_meta_format}
\title{format and reduce a list of variable metadata}
\usage{
var_meta_format(data)
}
\arguments{
\item{data}{a list of dataframes (with the expectation a single row per dataframe)}
}
\value{
a dataframe
}
\description{
Provided a list of variable meta data,  \code{var_meta_format} infers column types using \code{col_type} and converts
list and character variables that are not of the inferred type.
}
