% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables_to_dataframe}
\alias{variables_to_dataframe}
\title{Convert variable list to long-format data.frame}
\usage{
variables_to_dataframe(variables_list)
}
\arguments{
\item{variables_list}{List of variable objects from API response.}
}
\value{
A data.frame in long format.
}
\description{
Convert variable list to long-format data.frame
}
\keyword{internal}
