% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabularies.R
\name{vocabularies}
\alias{vocabularies}
\title{Controlled Vocabulary Functions}
\description{
Load and manage controlled vocabularies (CESSDA topics, SSJDA topics,
DDI sampling procedures, DDI modes of collection).
}
